Classes

+++ platforms/cpu/kernels/backward.hpp
@@ -14,8 +14,8 @@
 namespace Mixtape {
 
 template <typename REAL>
-void backward(const float* __restrict__ log_transmat,
-              const float* __restrict__ log_startprob,
+void backward(const double* __restrict__ log_transmat,
+              const double* __restrict__ log_startprob,
               const float* __restrict__ frame_logprob,
               const int sequence_length,
               const int n_states,
